refactor: add child windows property to state
