+++ src/parent/reducers/reducers.prod.js
@@ -1,2 +1,1 @@
-
 export default (reducer) => reducer;
